+++ R/0_imports.R
@@ -1,13 +1,13 @@
 #' @importFrom recipes add_step bake check_type ellipse_check is_trained names0
 #' @importFrom recipes prep printer rand_id sel2char step recipes_eval_select
 #' @importFrom recipes get_keep_original_cols tidy
-#' @importFrom rlang enquos abort
-#' @importFrom dplyr bind_cols bind_rows
+#' @importFrom rlang enquos abort quos_auto_name
+#' @importFrom dplyr bind_cols bind_rows left_join
 #' @importFrom tibble as_tibble tibble
 #' @importFrom tidyr pivot_longer
 #' @importFrom splines ns
 #' @importFrom generics tidy
-#' @importFrom stats coefficients predict na.omit lm formula
+#' @importFrom stats coefficients predict na.omit lm formula nextn
 #' @importFrom fftw IFFT FFT
 #' @importFrom utils data
 NULL